V1

+++ app/Http/Controllers/AgenceController.php
@@ -6,6 +6,13 @@ use Illuminate\Http\Request;
 use App\Models\Contrat;
 use App\Models\Agence;
 use App\Models\Role;
+use App\Models\Product;
+use App\Models\User;
+use App\Models\Customer;
+use App\Models\Employe;
+use App\Models\Facture;
+use App\Models\AgenceUser;
+use App\Models\ContratsEmployes;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
@@ -58,12 +65,46 @@ class AgenceController extends Controller
             return response()->json(['message' => 'Agence enregistré avec succès', 'status' => 'success']);
         }
 
+
     public function delete(Request $request)
         {
-
             Auth::user()->access('SUPPRESSION AGENCE');
 
             $agence = Agence::find($request->id);
+            $hasContrats = Contrat::where('agence_id', $agence->id)->exists();
+            $hasEmployes = Employe::where('agence_id', $agence->id)->exists();
+            $hasFactures = Facture::where('agence_id', $agence->id)->exists();
+            $hasAgenceUser = AgenceUser::where('agence_id', $agence->id)->exists();
+
+
+            if ($hasContrats) {
+                return response()->json([
+                    'message' => 'Impossible de supprimer cet élément : il est lié à un ou plusieurs contrats.',
+                    'status' => 'error'
+                ]);
+            }
+
+            if ($hasEmployes) {
+                return response()->json([
+                    'message' => 'Impossible de supprimer cet élément : il est lié à un ou plusieurs employés.',
+                    'status' => 'error'
+                ]);
+            }
+
+            if ($hasFactures) {
+                return response()->json([
+                    'message' => 'Impossible de supprimer cet élément : il est lié à une ou plusieurs factures.',
+                    'status' => 'error'
+                ]);
+            }
+
+            if ($hasAgenceUser) {
+                return response()->json([
+                    'message' => 'Impossible de supprimer cet élément : il est lié à un ou plusieurs agent.',
+                    'status' => 'error'
+                ]);
+            }
+            
 
             if($agence->delete()){
                 return response()->json(['message' => 'Agence supprimé avec succès',"status"=>"success"]);
