+++ app/Http/Controllers/CustomerController.php
@@ -6,6 +6,8 @@
     use App\Models\Customer;
     use App\Models\Role;
     use App\Models\User;
+    use App\Models\Facture;
+    use App\Models\Contrat;
     use Illuminate\Support\Facades\Auth;
     use Illuminate\Support\Facades\Storage; 
 
@@ -100,6 +102,23 @@
             Auth::user()->access('SUPPRESSION CLIENT');
 
             $customer = Customer::find($request->id);
+            $hasContrats = Contrat::where('customer_id', $customer->id)->exists();
+            $hasFactures = Facture::where('customer_id', $customer->id)->exists();
+
+
+            if ($hasContrats) {
+                return response()->json([
+                    'message' => 'Impossible de supprimer cet élément : il est lié à un ou plusieurs contrats.',
+                    'status' => 'error'
+                ]);
+            }
+
+            if ($hasFactures) {
+                return response()->json([
+                    'message' => 'Impossible de supprimer cet élément : il est lié à une ou plusieurs factures.',
+                    'status' => 'error'
+                ]);
+            }
 
             if($customer->delete()){
                 return response()->json(['message' => 'Client supprimé avec succès',"status"=>"success"]);
